pyproject.toml
[tool.poetry]
name = "cardian"
version = "0.1.0"
description = ""
authors = ["artem <sutalkuar808typ@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
scikit-learn = "^1.8.0"
pandas = "^3.0.2"
numpy = "^2.4.4"
matplotlib = "^3.10.8"
fastapi = "^0.135.3"
uvicorn = "^0.42.0"
sqlalchemy = "^2.0.49"
alembic = "^1.18.4"
asyncpg = "^0.31.0"
psycopg = {extras = ["binary"], version = "^3.3.3"}
pydantic-settings = "^2.13.1"
python-dotenv = "^1.2.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.12.1"
python-multipart = "^0.0.22"
pydantic = "^2.12.5"
email-validator = "^2.3.0"
aiogram = "^3.26.0"
aiohttp = "^3.13.5"
aiofiles = "^25.1.0"
aiosmtplib = "^5.1.0"
celery = "^5.6.3"
redis = "^7.4.0"
flower = "^2.0.1"
docker = "^7.1.0"
xgboost = "^3.2.0"
catboost = "^1.2.10"
google-genai = "^1.73.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
ruff = "^0.15.9"
black = "^26.3.1"
mypy = "^1.20.0"
httpx = "^0.28.1"
locust = "^2.43.4"
ipykernel = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
